Fix up a few more project.json files for core-setup.
Building a CoreCLR with a package version that doesn't exist anywhere I
uncovered two more project.json files that need to be updated in order to
build core-setup.  One simply needs the runtimes list to be filtered and
the other needs a runtimes section to be added; for this I added a new
method, AddRuntimeId, to the UpdateProjectJson class.